+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.12.1](https://github.com/chanzuckerberg/single-cell-explorer/compare/v1.12.0...v1.12.1) (2025-07-09)
+
+
+### Bug Fixes
+
+* GeneSets bug ([#1232](https://github.com/chanzuckerberg/single-cell-explorer/issues/1232)) ([43780a5](https://github.com/chanzuckerberg/single-cell-explorer/commit/43780a5705d0064a14bc76d09a9ed3575fa79c33))
+* skip truncate test ([#1235](https://github.com/chanzuckerberg/single-cell-explorer/issues/1235)) ([60f6cea](https://github.com/chanzuckerberg/single-cell-explorer/commit/60f6ceabeb7b9e9fbd1dd8d34426a937f65b2744))
+
+## [1.12.0](https://github.com/chanzuckerberg/single-cell-explorer/compare/v1.11.1...v1.12.0) (2025-07-04)
+
+
+### Features
+
+* Update Explorer to support index by feature_id ([#1222](https://github.com/chanzuckerberg/single-cell-explorer/issues/1222)) ([3daf593](https://github.com/chanzuckerberg/single-cell-explorer/commit/3daf59368b0a2bf0bc2d8c107d0f194354a9dc5d))
+* Updates to Gene panel readability ([#1230](https://github.com/chanzuckerberg/single-cell-explorer/issues/1230)) ([e7f69be](https://github.com/chanzuckerberg/single-cell-explorer/commit/e7f69be123bdf44d6cf4aa543ebce2425ebab156))
+
+
+### Bug Fixes
+
+* Address UI feedback ([#1229](https://github.com/chanzuckerberg/single-cell-explorer/issues/1229)) ([1dfd601](https://github.com/chanzuckerberg/single-cell-explorer/commit/1dfd6017671dc9763b5fe9c1d75bcfa61654957f))
+
 ## [1.11.1](https://github.com/chanzuckerberg/single-cell-explorer/compare/v1.11.0...v1.11.1) (2025-06-18)